+++ nemo_curator/eval/llm_judge/__init__.py
@@ -12,6 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .llm_judge_workflow import LLMJudgeWorkflow
+from .workflow import LLMJudgeWorkflow
 
 __all__ = ["LLMJudgeWorkflow"]
